de-spaghettify, fix some spelling, update docs, other fixes
