--- megatron/neox_arguments/tokenizer.py
@@ -1,10 +1,0 @@
-from dataclasses import dataclass
-
-@dataclass
-class NeoXArgsTokenizer:
-
-    tokenizer_type: str = "GPT2BPETokenizer"
-    """"""
-
-    padded_vocab_size: int = None
-    """"""
